git-svn-id: https://svn.code.sf.net/p/numere/source/trunk@1130 d50cd26f-24d5-421e-9e7d-30b8b4216d91

+++ gui/editor/editor.cpp
@@ -2736,9 +2736,9 @@ static bool isOnlyActiveBlock(int nBlockID, std::vector<int> vBlocks)
 {
     for (size_t i = 0; i < vBlocks.size(); i++)
     {
-        if (i == nBlockID && vBlocks[i] != 1)
+        if ((int)i == nBlockID && vBlocks[i] != 1)
             return false;
-        else if (i != nBlockID && vBlocks[i])
+        else if ((int)i != nBlockID && vBlocks[i])
             return false;
     }
